Initial support for multiple bot instances This allows to run multiple bot instances on the same app.
Right now it is not possible to select which plugins will be available
in each instance, but I'm planning to implement that soon. The idea so
far is to keep the plugins registrations on the init() func but make
possible to disable a command in an instance.

Also I don't think that the PeriodicCommands must be registered
globally, as at least for my needs I have different channels and needs
for each instance, so it would be best to register this kind of commands
in a per instance basis.