Allow Protocol Argument Passing
Due to the generic nature of this framework, platform specific
behaviors are hard to take advantage of. With this PR, protocol
plugins can offer additional information for message handlers to
read, and accept additional information as well.

In particular, this change was tested to allow a Slack response to
be made as a threaded reply, instead of only ever in the main channel.

EG:
func threadedEcho(cmd *bot.Cmd) (bot.CmdResult, error) {
	var params *slack.PostMessageParameters
	if ev, ok := cmd.MessageData.ProtoMsg.(*slack.MessageEvent); ok {
		params = &slack.PostMessageParameters{
			AsUser:          true,
			ThreadTimestamp: ev.Timestamp,
		}
	}
	return bot.CmdResult{
		Message:     fmt.Sprintf("%s: %s", cmd.User.Nick, cmd.Raw),
		Channel:     cmd.Channel,
		ProtoParams: params,
	}, nil
}