Slack: Add methods to directly add or remove reactions
These methods are simple helper methods to expose Slack emoji
reactions to bot commands. Any command that wants to call these
methods obviously is responsible for ensuring that their command
is only used in Slack use cases.

My favorite use case for this is to have the bot auto react to usages
of @here with :please-no-at-here:.

Future Plans: Build a full Reaction Router pattern so that the bots can
register commands to run when a message is tagged with a reaction.
For example:
https://reacji-channeler.builtbyslack.com/