Use SendMessageV2 everywhere (#116)
Then rename SendMessageV2 to SendMessage because it does the same thing.
This is a breaking change, but it is an easy one to fix, so I want to do
it anyway.

This isn't necessary everywhere, but for standardization, I thought
it would be good to really deprecate SendMessageV1.

The place this IS needed is in PeriodicCommands v2 and PassiveCommands
V2. In both those places, objects come back from the caller with ProtoParams
that should be forwarded through to the message sending.